Aligned bytecode to 4 bytes

+++ src/processor/control_unit/instruction.rs
@@ -10,43 +10,43 @@ pub enum ComparisonType {
 }
 
 pub struct MoveInstruction {
-    pub destination_register: u8,
+    pub destination_register: u32,
     pub value: Operand,
 }
 
 pub struct AddInstruction {
-    pub destination_register: u8,
+    pub destination_register: u32,
     pub first_operand: Operand,
     pub second_operand: Operand,
 }
 
 pub struct SubInstruction {
-    pub destination_register: u8,
+    pub destination_register: u32,
     pub first_operand: Operand,
     pub second_operand: Operand,
 }
 
 pub struct MulInstruction {
-    pub destination_register: u8,
+    pub destination_register: u32,
     pub first_operand: Operand,
     pub second_operand: Operand,
 }
 
 pub struct DivInstruction {
-    pub destination_register: u8,
+    pub destination_register: u32,
     pub first_operand: Operand,
     pub second_operand: Operand,
 }
 
 pub struct SimilarityInstruction {
-    pub destination_register: u8,
+    pub destination_register: u32,
     pub first_operand: Operand,
     pub second_operand: Operand,
 }
 
 pub struct JumpCompareInstruction {
     pub comparison_type: ComparisonType,
-    pub bytecode_jump_index: u8,
+    pub bytecode_jump_index: u32,
     pub first_operand: Operand,
     pub second_operand: Operand,
 }
@@ -56,7 +56,7 @@ pub struct OutputInstruction {
 }
 
 pub struct LoadInstruction {
-    pub destination_register: u8,
+    pub destination_register: u32,
     pub file_path: String,
 }
