+++ src/processor/control_unit/semantic_logic_unit/value.rs
@@ -1,6 +1,6 @@
 #[derive(Clone, Debug, PartialEq)]
 pub enum Value {
     Text(String),
-    Number(u8),
+    Number(u32),
     None,
 }
